Initial major refactor pass
